# server/src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserRole": {
				"type": "string",
				"enum": [
					"ADMIN",
					"PROJECT_MANAGER",
					"COLLABORATOR"
				]
			},
			"User": {
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"mustResetPassword": {
						"type": "boolean"
					},
					"createdAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"updatedAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					}
				},
				"required": [
					"id",
					"email",
					"name",
					"role",
					"mustResetPassword",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserProfile": {
				"$ref": "#/components/schemas/User"
			},
			"ApiResponse_UserProfile_": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"success",
							"error"
						]
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserProfile"
							}
						],
						"nullable": true
					},
					"error": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"data",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateProfileRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_null_": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"success",
							"error"
						]
					},
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [],
						"nullable": true
					},
					"error": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"data",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChangePasswordRequest": {
				"properties": {
					"currentPassword": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					},
					"confirmNewPassword": {
						"type": "string"
					}
				},
				"required": [
					"currentPassword",
					"newPassword",
					"confirmNewPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_36_Enums.Priority": {
				"type": "string",
				"enum": [
					"LOW",
					"MEDIUM",
					"HIGH"
				]
			},
			"_36_Enums.Status": {
				"type": "string",
				"enum": [
					"TODO",
					"IN_PROGRESS",
					"COMPLETED"
				]
			},
			"DefaultSelection_Prisma._36_TaskPayload_": {
				"properties": {
					"assignedUserId": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/_36_Enums.Status"
					},
					"priority": {
						"$ref": "#/components/schemas/_36_Enums.Priority"
					},
					"dueDate": {
						"type": "string",
						"format": "date-time"
					},
					"description": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"assignedUserId",
					"updatedAt",
					"createdAt",
					"status",
					"priority",
					"dueDate",
					"description",
					"title",
					"id"
				],
				"type": "object"
			},
			"Task": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_TaskPayload_",
				"description": "Model Task"
			},
			"ApiResponse_Task-Array_": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"success",
							"error"
						]
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Task"
						},
						"type": "array",
						"nullable": true
					},
					"error": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"data",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Task_": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"success",
							"error"
						]
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Task"
							}
						],
						"nullable": true
					},
					"error": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"data",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Priority": {
				"$ref": "#/components/schemas/_36_Enums.Priority"
			},
			"Status": {
				"$ref": "#/components/schemas/_36_Enums.Status"
			},
			"CreateTaskInput": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"assignedUserId": {
						"type": "string"
					},
					"dueDate": {
						"type": "string",
						"format": "date-time"
					},
					"priority": {
						"$ref": "#/components/schemas/Priority"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					}
				},
				"required": [
					"title"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateTaskInput": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"assignedUserId": {
						"type": "string"
					},
					"dueDate": {
						"type": "string",
						"format": "date-time"
					},
					"priority": {
						"$ref": "#/components/schemas/Priority"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse__publicKey-string__": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"success",
							"error"
						]
					},
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"publicKey": {
								"type": "string"
							}
						},
						"required": [
							"publicKey"
						],
						"type": "object",
						"nullable": true
					},
					"error": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"data",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PushSubscriptionKeys": {
				"properties": {
					"p256dh": {
						"type": "string"
					},
					"auth": {
						"type": "string"
					}
				},
				"required": [
					"p256dh",
					"auth"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PushSubscriptionRequest": {
				"properties": {
					"endpoint": {
						"type": "string"
					},
					"keys": {
						"$ref": "#/components/schemas/PushSubscriptionKeys"
					}
				},
				"required": [
					"endpoint",
					"keys"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PingResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"time": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"message",
					"time"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Attachment": {
				"properties": {
					"id": {
						"type": "string"
					},
					"filename": {
						"type": "string"
					},
					"fileUrl": {
						"type": "string"
					},
					"mimeType": {
						"type": "string"
					},
					"fileSize": {
						"type": "number",
						"format": "double"
					},
					"taskId": {
						"type": "string"
					},
					"commentId": {
						"type": "string",
						"nullable": true
					},
					"uploadedById": {
						"type": "string"
					},
					"createdAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					}
				},
				"required": [
					"id",
					"filename",
					"fileUrl",
					"mimeType",
					"fileSize",
					"taskId",
					"commentId",
					"uploadedById",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Comment": {
				"properties": {
					"id": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"taskId": {
						"type": "string"
					},
					"authorId": {
						"type": "string"
					},
					"author": {
						"$ref": "#/components/schemas/User"
					},
					"createdAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"updatedAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"attachments": {
						"items": {
							"$ref": "#/components/schemas/Attachment"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"content",
					"taskId",
					"authorId",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Comment-Array_": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"success",
							"error"
						]
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Comment"
						},
						"type": "array",
						"nullable": true
					},
					"error": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"data",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Comment_": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"success",
							"error"
						]
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Comment"
							}
						],
						"nullable": true
					},
					"error": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"data",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateAttachmentRequest": {
				"properties": {
					"filename": {
						"type": "string"
					},
					"fileUrl": {
						"type": "string"
					},
					"mimeType": {
						"type": "string"
					},
					"fileSize": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"filename",
					"fileUrl",
					"mimeType",
					"fileSize"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCommentRequest": {
				"properties": {
					"content": {
						"type": "string"
					},
					"attachments": {
						"items": {
							"$ref": "#/components/schemas/CreateAttachmentRequest"
						},
						"type": "array"
					}
				},
				"required": [
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthData": {
				"properties": {
					"token": {
						"type": "string"
					},
					"mustResetPassword": {
						"type": "boolean"
					}
				},
				"required": [
					"token",
					"mustResetPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_AuthData_": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"success",
							"error"
						]
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/AuthData"
							}
						],
						"nullable": true
					},
					"error": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"data",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResetPasswordRequest": {
				"properties": {
					"currentPassword": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					},
					"confirmNewPassword": {
						"type": "string"
					}
				},
				"required": [
					"currentPassword",
					"newPassword",
					"confirmNewPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Attachment-Array_": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"success",
							"error"
						]
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Attachment"
						},
						"type": "array",
						"nullable": true
					},
					"error": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"data",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Attachment_": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"success",
							"error"
						]
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Attachment"
							}
						],
						"nullable": true
					},
					"error": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"data",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "@nextask/server",
		"version": "1.0.0",
		"contact": {}
	},
	"paths": {
		"/users/me": {
			"get": {
				"operationId": "GetMe",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserProfile_"
								}
							}
						}
					}
				},
				"description": "Returns the authenticated user's profile.",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"patch": {
				"operationId": "UpdateMe",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserProfile_"
								}
							}
						}
					}
				},
				"description": "Updates the authenticated user's name and/or email.",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProfileRequest"
							}
						}
					}
				}
			}
		},
		"/users/me/change-password": {
			"post": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "Allows an authenticated, onboarded user to change their password voluntarily.\r\n(First-login forced reset uses POST /auth/reset-password instead.)",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangePasswordRequest"
							}
						}
					}
				}
			}
		},
		"/tasks": {
			"get": {
				"operationId": "GetTasks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Task-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateNewTask",
				"responses": {
					"201": {
						"description": "Task Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Task_"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTaskInput"
							}
						}
					}
				}
			}
		},
		"/tasks/{id}": {
			"get": {
				"operationId": "GetTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Task_"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateExistingTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Task_"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTaskInput"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteExistingTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/push/key": {
			"get": {
				"operationId": "GetPublicKey",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse__publicKey-string__"
								}
							}
						}
					}
				},
				"description": "Retrieves the VAPID Public Key needed to subscribe to push notifications.",
				"tags": [
					"Push Notifications"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/push/subscribe": {
			"post": {
				"operationId": "Subscribe",
				"responses": {
					"201": {
						"description": "Subscribed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "Subscribes the current user's device/browser to push notifications.",
				"tags": [
					"Push Notifications"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PushSubscriptionRequest"
							}
						}
					}
				}
			}
		},
		"/push/unsubscribe": {
			"post": {
				"operationId": "Unsubscribe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "Unsubscribes the current user's device/browser from push notifications.",
				"tags": [
					"Push Notifications"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"endpoint": {
										"type": "string"
									}
								},
								"required": [
									"endpoint"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/ping": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PingResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/tasks/{taskId}/comments": {
			"get": {
				"operationId": "GetComments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Comment-Array_"
								}
							}
						}
					}
				},
				"description": "Retrieves all comments for a specific task.",
				"tags": [
					"Comments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateComment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Comment_"
								}
							}
						}
					}
				},
				"description": "Posts a new comment to a specific task.",
				"tags": [
					"Comments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCommentRequest"
							}
						}
					}
				}
			}
		},
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_AuthData_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterRequest"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_AuthData_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest"
							}
						}
					}
				}
			}
		},
		"/auth/reset-password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_AuthData_"
								}
							}
						}
					}
				},
				"description": "First-login password reset (and any subsequent voluntary reset).\r\nRequires a valid JWT — users with mustResetPassword=true are ONLY\r\nallowed to call this endpoint until they complete the reset.",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResetPasswordRequest"
							}
						}
					}
				}
			}
		},
		"/tasks/{taskId}/attachments": {
			"get": {
				"operationId": "GetAttachments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Attachment-Array_"
								}
							}
						}
					}
				},
				"description": "Retrieves all attachments for a specific task.",
				"tags": [
					"Attachments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "UploadAttachmentMetadata",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Attachment_"
								}
							}
						}
					}
				},
				"description": "Uploads file metadata directly to a task.",
				"tags": [
					"Attachments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAttachmentRequest"
							}
						}
					}
				}
			}
		},
		"/attachments/{attachmentId}": {
			"delete": {
				"operationId": "RemoveAttachment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "Deletes an attachment metadata record by ID.",
				"tags": [
					"Attachments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "attachmentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}